feat: input accepted, added as test
Excitement: the raw approach took a couple minutes.  The smarter
approach (operating on and splitting and combining) ranges would likely
get around this, but I'm still excited that there was a problem that
took some time.  In particular, I look forward to trying par_iter with
rayon in the naive solution.