skeleton(parsing): laid out early parsing code
Most notably: spent a fair bit of time comparing Logos crate options.
It is an attractive crate with a number of upsides (including nominal
performance).  However, it uses custom rolled regex, has lacking
documentation on a variety of issues (particularly related to the regex
implementation and limitations).

By constrat BurntSushi's regex crate is heavily used and the amount of
attention and their work (and writings on regex) instill a lot of faith
regarding correctness.  It's also a more reliable standard.

The default regex crate is balancing a number of issues that will impact
performance. (e.g. compiling regex at runtime and choosing regex
implementations to avoid pathological compile cases)

However, even if that were a forced cost the reliability and
standardization (and excellent documentation) would probably be a
worthwhile trade-off for most cases.  Additionally, however: BurntSushi
recently published a `regex-automata` crate allowing access to the
underlyign models.  Including dense DFA and on the fly calculated NFA
models.  For a known regex, compilation can be done ahead of time and
rolled into the crate via include_bytes! macro.

This feels like an excellent option, with long legs should I decide to
improve performance that way.